Update and move graphics integration test from drivers to graphics
